add text completion
