+++ evalscope/perf/plugin/api/__init__.py
@@ -1,4 +1,3 @@
 from evalscope.perf.plugin.api.custom_api import CustomPlugin
 from evalscope.perf.plugin.api.dashscope_api import DashScopeApiPlugin
-from evalscope.perf.plugin.api.local_infer_api import LocalPlugin
 from evalscope.perf.plugin.api.openai_api import OpenaiPlugin
