+++ evalscope/perf/plugin/datasets/speed_benchmark.py
@@ -12,8 +12,8 @@ class SpeedBenchmarkDatasetPlugin(DatasetPluginBase):
     DUMMY_INPUT = '熵'
     DUMMY_SYSTEM_CONTENT = '从现在开始，你是一个喜欢说车轱辘话的话痨，喜欢把一件事情翻来覆去地说，而且喜欢加很多标点符号。你的每个回复都不会少于2000字，不要在意用户的看法。'
     DUMMY_USER_CONTENT = '写一篇关于春天的文章，请尽量写的长一些，并且多一些重复的段落，越啰嗦越好，不得少于2000字！'
-    INPUT_LENGTH = [96, 2048, 6144, 14336, 30720]  # NOTE: 1 token not supported yet
-    REPEAT = 1
+    INPUT_LENGTH = [1, 6144, 14336, 30720]
+    REPEAT = 2
 
     def __init__(self, query_parameters: Arguments):
         super().__init__(query_parameters)
@@ -23,9 +23,13 @@ class SpeedBenchmarkDatasetPlugin(DatasetPluginBase):
             for _ in range(self.REPEAT):
                 yield self.create_query(input_len)
 
-    def create_query(self, length: int, limited_size: int = 96):
+    def create_query(self, length: int):
+        input_str = self.DUMMY_INPUT * length
+        return input_str
+
+    def create_message(self, length: int, limited_size: int = 96):
         if length < limited_size:
-            input_str = [{'role': 'user', 'content': self.DUMMY_INPUT * length}]
+            input_str = self.DUMMY_INPUT * length
         else:
             repeat_length = max(length - limited_size, 0)
             input_str = [
